Fixed alias command not working with a custom ASDF_DATA_DIR
Added an error in case you try to make an alias to a non existing version

README.md: Added command to create the adopt-openjdk-11.0 alias since that version doesn't exist normally with asdf-java

Simplified lint.yaml

Added functional test which also includes a test for a custom ASDF_DATA_DIR

+++ lib/commands/command.bash
@@ -27,6 +27,8 @@
 set -e
 
 shopt -s nullglob
+# shellcheck source=/dev/null
+source "${ASDF_DIR}/lib/utils.bash"
 
 usage() {
   echo "usage"
@@ -135,7 +137,7 @@ main() {
   plugin="$1"
   # alias="$2" -- not used directly
   version="$3"
-  install_dir="$ASDF_DIR/installs/$plugin"
+  install_dir="$(asdf_data_dir)/installs/$plugin"
 
   [ ! -d "$install_dir" ] && abort "The plugin '$plugin' could not be found"
   cd "$install_dir"
@@ -161,6 +163,9 @@ main() {
             # *) abort "Don't know how to automatically alias $2" ;;
         esac
       else
+        if [ ! -e "${3}" ]; then
+          abort "Version $3 could not be found"
+        fi
         echo "$2 => $3"
         ln -nsf "$3" "$2" && reshim_plugin "$@"
       fi
